Desikan-Killiany-Tourville Cortical Atlas

Coordinate data for the Desikan-Killany-Tourville Cortical atlas, with 31 regions in on the cortical surface of the brain.

dkt()

Value

A [ggseg.formats::ggseg_atlas] object (cortical).

References

Klein A, Tourville J (2012). 101 labeled brain images and a consistent human cortical labeling protocol. *Frontiers in Neuroscience*, 6:171. doi:10.3389/fnins.2012.00171

Examples

dkt()
#> $atlas
#> [1] "dkt"
#> 
#> $type
#> [1] "cortical"
#> 
#> $data
#> Simple feature collection with 83 features and 5 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 0 ymin: 0 xmax: 1390.913 ymax: 205.5564
#> CRS:           NA
#> # A tibble: 83 × 6
#>    hemi  side    region                label     roi                    geometry
#>  * <chr> <chr>   <chr>                 <chr>     <chr>            <MULTIPOLYGON>
#>  1 left  lateral caudal middle frontal lh_cauda… 0002  (((110.9037 186.7029, 93…
#>  2 left  lateral fusiform              lh_fusif… 0005  (((256.0259 44.7229, 256…
#>  3 left  lateral inferior parietal     lh_infer… 0006  (((215.7759 163.1805, 21…
#>  4 left  lateral inferior temporal     lh_infer… 0007  (((239.2787 59.31508, 22…
#>  5 left  lateral lateral occipital     lh_later… 0009  (((275.7996 118.0405, 28…
#>  6 left  lateral lateral orbitofrontal lh_later… 0010  (((2.28531 70.05087, 2.1…
#>  7 left  lateral middle temporal       lh_middl… 0013  (((230.5115 105.4085, 22…
#>  8 left  lateral pars opercularis      lh_parso… 0016  (((69.60376 127.8272, 55…
#>  9 left  lateral pars orbitalis        lh_parso… 0017  (((44.75661 64.90361, 41…
#> 10 left  lateral pars triangularis     lh_parst… 0018  (((51.00001 116.0256, 37…
#> # ℹ 73 more rows
#> 
#> $palette
#>  caudal anterior cingulate      caudal middle frontal 
#>                  "#7D64A0"                  "#641900" 
#>                     cuneus                 entorhinal 
#>                  "#DC1464"                  "#DC140A" 
#>                   fusiform          inferior parietal 
#>                  "#B4DC8C"                  "#DC3CDC" 
#>          inferior temporal          isthmus cingulate 
#>                  "#B42878"                  "#8C148C" 
#>          lateral occipital      lateral orbitofrontal 
#>                  "#141E8C"                  "#234B32" 
#>                    lingual       medial orbitofrontal 
#>                  "#E18C8C"                  "#C8234B" 
#>            middle temporal            parahippocampal 
#>                  "#A06432"                  "#14DC3C" 
#>                paracentral           pars opercularis 
#>                  "#3CDC3C"                  "#DCB48C" 
#>             pars orbitalis          pars triangularis 
#>                  "#146432"                  "#DC3C14" 
#>              pericalcarine                postcentral 
#>                  "#78643C"                  "#DC1414" 
#>        posterior cingulate                 precentral 
#>                  "#DCB4DC"                  "#3C14DC" 
#>                  precuneus rostral anterior cingulate 
#>                  "#A08CB4"                  "#50148C" 
#>     rostral middle frontal           superior frontal 
#>                  "#4B327D"                  "#14DCA0" 
#>          superior parietal          superior temporal 
#>                  "#14B48C"                  "#8CDCDC" 
#>              supramarginal        transverse temporal 
#>                  "#50A014"                  "#9696C8" 
#>                     insula 
#>                  "#FFC020" 
#> 
#> attr(,"class")
#> [1] "brain_atlas"